requirements.txt
# MedAI Platform v1.1 — Dependencies
# Install: pip install -r requirements.txt

# Core framework
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Document processing
docling>=2.0.0
pymupdf>=1.23.0
pillow>=10.0.0

# PII stripping
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0

# Vector store + embeddings
chromadb>=0.4.0
sentence-transformers>=2.7.0

# Database
sqlcipher3>=0.5.0; sys_platform != "win32"
# fallback for Windows or if sqlcipher not available:
# pysqlite3>=0.5.0

# Claude API + structured output
anthropic>=0.30.0
instructor>=1.3.0

# Google Gemini API
google-generativeai>=0.3.0

# Web fetching
requests>=2.31.0
beautifulsoup4>=4.12.0
biopython>=1.83

# Async
aiohttp>=3.9.0

# UI
streamlit>=1.32.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# Utilities
networkx>=3.2.0
loguru>=0.7.0
